+++ packages/image_picker/image_picker_android/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## 0.8.13+19
 
+* Updates plugin to use Android Photo Picker on API 36 and above.
+* Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
 * Bumps the androidx group across 10 directories with 1 update.
 
 ## 0.8.13+18